pass raw sql to celery task
