+++ tdrs-backend/tdpservice/search_indexes/tasks.py
@@ -2,10 +2,16 @@
 
 from __future__ import absolute_import
 import time
+import csv
+import gzip
+import os
 from tdpservice.users.models import User
 from celery import shared_task
 from tdpservice.core.utils import log
 import subprocess
+from tdpservice.data_files.s3_client import S3Client
+from django.conf import settings
+from django.apps import apps
 
 
 def prettify_time_delta(start, end):
@@ -66,6 +72,77 @@ def reindex_elastic_documents():
 
 
 @shared_task
-def upload_local_file_to_s3(local_filename, s3_filename):
-    """Export a selected queryset to a csv file stored in s3."""
-    pass
+def export_queryset_to_s3_csv(query_str, query_params, field_names, model_name, s3_filename):
+    """
+    Export a selected queryset to a csv file stored in s3.
+
+    @param query_str: a sql string obtained via queryset.query.
+    @param s3_filename: a string representing the file path/name in s3.
+    """
+    class Echo:
+        """An object that implements just the write method of the file-like interface."""
+
+        def write(self, value):
+            """Write the value by returning it, instead of storing in a buffer."""
+            return value
+
+    class RowIterator:
+        """Iterator class to support custom CSV row generation."""
+
+        def __init__(self, field_names, queryset):
+            self.field_names = field_names
+            self.queryset = queryset
+            self.writer = csv.writer(Echo())
+            self.is_header_row = True
+            self.header_row = self.__init_header_row(field_names)
+
+        def __init_header_row(self, field_names):
+            """Generate custom header row."""
+            header_row = []
+            for name in field_names:
+                header_row.append(name)
+                if name == "datafile":
+                    header_row.append("STT")
+            return header_row
+
+        def __iter__(self):
+            """Yield the next row in the csv export."""
+            for obj in self.queryset.iterator():
+                # for obj in self.queryset:
+                row = []
+
+                if self.is_header_row:
+                    self.is_header_row = False
+                    yield self.writer.writerow(self.header_row)
+
+                for field_name in self.field_names:
+                    field = getattr(obj, field_name)
+                    row.append(field)
+                    if field and field_name == "datafile":
+                        # print(field)
+                        row.append(field.stt.stt_code)
+                yield self.writer.writerow(row)
+
+    Model = apps.get_model('search_indexes', model_name)
+    queryset = Model.objects.raw(query_str, query_params)
+    iterator = RowIterator(field_names, queryset)
+    s3 = S3Client()
+
+    # stream the creation of the file using smart_open
+    # url = f's3://{settings.AWS_S3_DATAFILES_BUCKET_NAME}/{s3_filename}.csv'
+    # with open(url, 'w', transport_params={'client': s3.client}) as fout:
+    #     for _, s in enumerate(iterator):
+    #         fout.write(s)
+
+    # write + compress into tar/gz on filesystem - upload resulting file to s3
+    # have to clean up resulting file
+
+    tmp_filename = 'file.csv.gz'
+    with gzip.open(tmp_filename, 'wt') as f:
+        for _, s in enumerate(iterator):
+            f.write(s)
+
+    local_filename = os.path.basename(tmp_filename)
+    s3.client.upload_file(local_filename, settings.AWS_S3_DATAFILES_BUCKET_NAME, s3_filename)
+
+    os.remove(local_filename)
